Use existing ConfigurationLatticeRandom and move it to core
