+++ etomica-core/src/main/java/etomica/config/ConfigurationLatticeRandom.java
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package etomica.modules.chainequilibrium;
+package etomica.config;
 
 import etomica.box.Box;
 import etomica.config.ConfigurationLattice;
