+++ etomica-modules/src/main/java/etomica/modules/chainequilibrium/ChainEquilibriumSim.java
@@ -12,6 +12,7 @@ import etomica.atom.AtomType;
 import etomica.atom.IAtom;
 import etomica.atom.IAtomList;
 import etomica.box.Box;
+import etomica.config.ConfigurationLatticeRandom;
 import etomica.data.meter.MeterTemperature;
 import etomica.integrator.IntegratorHard;
 import etomica.integrator.IntegratorMD.ThermostatType;
